restructure the registrations custom app, relocated notification template, and update cms commit tag
