+++ apcd-cms/src/apps/registrations/templates/submission_form/submission_error.html
@@ -1,0 +1,18 @@
+{% extends "base.html" %}
+{% block content %}
+<div class="container">
+    <div class="row">
+        <div class="col">
+            <h1>Request to Submit</h1>
+            <hr style="margin-top: 10px" />
+        </div>
+    </div>
+    <div class="c-message c-message--error">
+        There was an error when attempting to store your registration!
+        Please submit a ticket.
+    </div>
+    <form class="button-wrapper submit" action="/workbench/dashboard" method="GET">
+        <button class="form-button" type="submit">Go to Dashboard</button>
+    </form>
+</div>
+{% endblock %}
