+++ apcd-cms/src/taccsite_cms/custom_app_settings.py
@@ -1,3 +1,3 @@
-CUSTOM_APPS = ['apps.admin_regis_table', 'apps.submission_form']
+CUSTOM_APPS = ['apps.admin_regis_table', 'apps.registrations']
 CUSTOM_MIDDLEWARE = []
 STATICFILES_DIRS = ('taccsite_custom/apcd-cms',)
