+++ apcd-cms/src/taccsite_cms/urls_custom.py
@@ -2,5 +2,5 @@ from django.urls import path, include
 
 custom_urls = [
     path('administration/', include('apps.admin_regis_table.urls', namespace='administration')),
-    path('register/', include('apps.submission_form.urls', namespace='apcd'))
+    path('register/', include('apps.registrations.urls', namespace='register'))
 ]
